# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HindaviSwarajya Seva Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
  /users/{id}/follow:
    post:
      operationId: toggleFollow
      tags: [users]
      summary: Follow or unfollow a user (auth required)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Follow status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResult"
        "401":
          description: Not authenticated
  /users/{id}/posts:
    get:
      operationId: getUserPosts
      tags: [users]
      summary: Get posts by a specific user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User's seva posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SevaPost"
...
components:
  schemas:
...
    Rank:
      type: string
      enum:
        - Sevak
        - Mavla
        - Bargir
        - Shiledar
        - Naik
        - Havaldar
        - Jumledar
        - Hazari
        - Panch Hazari
        - Sardar
        - Killedar
        - Deshpande
        - Deshmukh
        - Subhedar
        - Ashta Pradhan
        - Sarnobat
        - Sar Senapati
    SevaCategory:
      type: string
      enum: [Food, Education, Health, Shelter, Other]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        avatar:
          type: string
        location:
          type: string
        rank:
          $ref: "#/components/schemas/Rank"
        totalHelped:
          type: integer
        mudra:
          type: integer
          description: Seva points — totalHelped × 10
        chhava:
          type: boolean
          description: Honorary rank awarded by admins; displayed above the normal rank
        followersCount:
          type: integer
        postsCount:
          type: integer
      required:
        - id
        - name
        - avatar
        - location
        - rank
        - totalHelped
        - mudra
        - chhava
        - followersCount
        - postsCount
    UserProfile:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            bio:
              type: string
            joinedAt:
              type: string
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userAvatar:
          type: string
        content:
          type: string
        timestamp:
          type: string
      required:
        - id
        - postId
        - userId
        - userName
        - userAvatar
        - content
        - timestamp
    SevaPost:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        content:
          type: string
        category:
          $ref: "#/components/schemas/SevaCategory"
        helpedPeople:
          type: integer
        likes:
          type: integer
        likedBy:
          type: array
          items:
            type: integer
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        tags:
          type: array
          items:
            type: string
        image:
          type: string
          nullable: true
        timestamp:
          type: string
        location:
          type: string
          nullable: true
        approvalStatus:
          type: string
          enum: [pending, approved, rejected]
      required:
        - id
        - userId
        - user
        - content
        - category
        - helpedPeople
        - likes
        - likedBy
        - comments
        - tags
        - timestamp
...
    FollowResult:
      type: object
      properties:
        following:
          type: boolean
        followersCount:
          type: integer
      required:
        - following
        - followersCount
...